Remove redundant uint32(id) cast

+++ contracts/core/accounts/facets/AccountsUpdateStatusEndowments.sol
@@ -75,7 +75,7 @@ contract AccountsUpdateStatusEndowments is
     state.STATES[id].closingEndowment = true;
     state.STATES[id].closingBeneficiary = beneficiary;
 
-    require(checkFullyExited(uint32(id)), "Not fully exited");
+    require(checkFullyExited(id), "Not fully exited");
     state.ENDOWMENTS[id] = tempEndowment;
     emit EndowmentUpdated(id);
     // emit EndowmentStateUpdated(id);
